Re-add images not correctly merged from djp branch, so that the forum view works again. Remove a reference to an icon which depended on isAdmin, which is only on djp branch.
Make extension profile use the correct name Ploneboard, not PloneBoard, and
make it set up the tool portal_ploneboard, not ploneboard_tool, as the rest
of the code and template expect. Note: Installing plone with the Ploneboard
extensino profile does not work, since the workflows are not applied!

Make PloneboardConversation implement canSetDefaultPage to return False, it
makes no sense to set a single comment as the default page of a conversation.

Make the "powered by" footer actually display the link by not using an ignored
tal: construct.

At least it runs now (Plone 2.5/Zope 2.9); note that moderation (e.g. the 
'retract' button) is broken because moderate.js relies on a deprecated-and-
disabled vcXMLRPC.js.